backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
google-auth==2.37.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pip==23.0.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyotp==2.9.0
requests==2.32.3
rsa==4.9
setuptools==65.5.1
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
django-extensions
gunicorn
django-cors-headers

